// bidly/apps/server/src/app/open-api/pre-standard/types/pre-standard-foreign.type.ts
/**
 * 사전규격 외자
 * @author 최강훈 <ganghun@lepisode.team>
 */
export type OpenAPIPreStandardForeign = {
  bsnsDivNm: string;
  refNo: string;
  prdctClsfcNoNm: string;
  orderInsttNm: string;
  rlDminsttNm: string;
  asignBdgtAmt: string;
  rcptDt: string;
  opninRgstClseDt: string;
  ofclTelNo: string;
  ofclNm: string;
  swBizObjYn: string;
  dlvrTmlmtDt: string;
  dlvrDaynum: string;
  bfSpecRgstNo: string;
  specDocFileUrl1: string;
  specDocFileUrl2: string;
  specDocFileUrl3: string;
  specDocFileUrl4: string;
  specDocFileUrl5: string;
  rgstDt: string;
  chgDt: string;
  bidNtceNoList: string;
};